Tutor/backend/requirements.txt
# TutorGPT Backend Dependencies
# Install with: pip install -r requirements.txt

# Core Framework
fastapi==0.121.1
uvicorn[standard]==0.38.0

# Agent Framework
openai-agents==0.5.0
openai==2.7.1

# AI/ML
google-genai==1.49.0
chromadb==1.3.4

# Database
aiosqlite==0.21.0
sqlalchemy==2.0.36

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
pydantic==2.12.4
pydantic-settings==2.11.0
python-dotenv==1.2.1
sentence-transformers==3.3.1

# Testing
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
